Also move parse_content() and save_hashtags() to Post, fix api
